Created folder "src/util". Put files appInterface.js, tinymceInit.js, i18n.js, helper.js, appFactory.js Include the associated refactors Include a bugfix in appInterface.js Include a beginning of implementation of live message
