+++ frontend/src/action-creator.sync.js
@@ -91,6 +91,7 @@ export const setWorkspaceMemberList = workspaceMemberList => ({ type: `${SET}/${
 export const WORKSPACE_MEMBER_ADD = `${WORKSPACE_MEMBER}/${ADD}`
 export const WORKSPACE_MEMBER_REMOVE = `${WORKSPACE_MEMBER}/${REMOVE}`
 export const removeWorkspaceMember = memberId => ({ type: `${REMOVE}/${WORKSPACE_MEMBER}`, memberId })
+export const addWorkspaceMember = member => ({ type: `${ADD}/${WORKSPACE_MEMBER}`, member })
 
 export const WORKSPACE_RECENT_ACTIVITY = `${WORKSPACE}/RecentActivity/List`
 export const WORKSPACE_RECENT_ACTIVITY_LIST = `${WORKSPACE_RECENT_ACTIVITY}/List`
