+++ frontend/src/component/Dashboard/RecentActivity.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import classnames from 'classnames'
 import { Link } from 'react-router-dom'
 import { Badge, ROLE } from 'tracim_frontend_lib'
-import { PAGE } from '../../helper.js'
+import { PAGE } from '../../util/helper.js'
 
 require('./RecentActivity.styl')
