+++ frontend/src/component/Header/Logo.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import { Link } from 'react-router-dom'
-import { PAGE } from '../../helper.js'
+import { PAGE } from '../../util/helper.js'
 
 const Logo = props => {
   return (
