+++ frontend/src/component/Header/MenuActionListItem/MenuProfil.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
 import PropTypes from 'prop-types'
-import { PAGE } from '../../../helper.js'
+import { PAGE } from '../../../util/helper.js'
 import { translate } from 'react-i18next'
 import { Avatar } from 'tracim_frontend_lib'
