+++ frontend/src/component/Sidebar/WorkspaceListItem.jsx
@@ -5,7 +5,7 @@ import { translate } from 'react-i18next'
 import PropTypes from 'prop-types'
 import AnimateHeight from 'react-animate-height'
 import { DropTarget } from 'react-dnd'
-import { DRAG_AND_DROP } from '../../helper.js'
+import { DRAG_AND_DROP } from '../../util/helper.js'
 import { ROLE } from 'tracim_frontend_lib'
 
 const qs = require('query-string')
