+++ frontend/src/component/Workspace/OpenCreateContentApp.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { connect } from 'react-redux'
 import { withRouter } from 'react-router'
-import appFactory from '../../appFactory.js'
+import appFactory from '../../util/appFactory.js'
 import { HACK_COLLABORA_CONTENT_TYPE } from '../../container/WorkspaceContent.jsx'
 
 const qs = require('query-string')
