+++ frontend/src/container/AdminAccount.jsx
@@ -33,7 +33,7 @@ import {
 import {
   editableUserAuthTypeList,
   PAGE
-} from '../helper.js'
+} from '../util/helper.js'
 import AgendaInfo from '../component/Dashboard/AgendaInfo.jsx'
 
 class Account extends React.Component {
