+++ frontend/src/container/ForgotPasswordNoEmailNotif.jsx
@@ -6,7 +6,7 @@ import CardHeader from '../component/common/Card/CardHeader.jsx'
 import CardBody from '../component/common/Card/CardBody.jsx'
 import FooterLogin from '../component/Login/FooterLogin.jsx'
 import Button from '../component/common/Input/Button.jsx'
-import { PAGE } from '../helper.js'
+import { PAGE } from '../util/helper.js'
 import { resetBreadcrumbs, setConfig } from '../action-creator.sync.js'
 import { CUSTOM_EVENT, buildHeadTitle } from 'tracim_frontend_lib'
 import { getConfig } from '../action-creator.async'
