+++ frontend/src/container/Header.jsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { connect } from 'react-redux'
 import { withRouter } from 'react-router'
-import i18n from '../i18n.js'
-import appFactory from '../appFactory.js'
+import i18n from '../util/i18n.js'
+import appFactory from '../util/appFactory.js'
 import { translate } from 'react-i18next'
 import * as Cookies from 'js-cookie'
 import Logo from '../component/Header/Logo.jsx'
@@ -27,7 +27,7 @@ import {
   PAGE,
   unLoggedAllowedPageList,
   ALL_CONTENT_TYPES
-} from '../helper.js'
+} from '../util/helper.js'
 import Search from '../component/Header/Search.jsx'
 import { Link } from 'react-router-dom'
 import {
