+++ frontend/src/container/Tracim.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { connect } from 'react-redux'
 import { translate } from 'react-i18next'
 import * as Cookies from 'js-cookie'
-import i18n from '../i18n.js'
+import i18n from '../util/i18n.js'
 import {
   Route, withRouter, Redirect
 } from 'react-router-dom'
@@ -20,16 +20,18 @@ import FlashMessage from '../component/FlashMessage.jsx'
 import WorkspaceContent from './WorkspaceContent.jsx'
 import Home from './Home.jsx'
 import WIPcomponent from './WIPcomponent.jsx'
+import { LiveMessageManager } from '../util/LiveMessageManager.js'
 import { buildHeadTitle, CUSTOM_EVENT, PROFILE } from 'tracim_frontend_lib'
 import {
   PAGE,
   COOKIE_FRONTEND,
   unLoggedAllowedPageList,
   getUserProfile
-} from '../helper.js'
+} from '../util/helper.js'
 import {
   getConfig,
   getAppList,
+  getLiveMessage,
   getContentTypeList,
   getUserIsConnected,
   getMyselfWorkspaceList,
@@ -56,6 +58,8 @@ export class Tracim extends React.Component {
   constructor (props) {
     super(props)
 
+    this.liveMessageManager = new LiveMessageManager()
+    this.liveMessageManager.openLiveMessageConnection()
     document.addEventListener(CUSTOM_EVENT.APP_CUSTOM_EVENT_LISTENER, this.customEventReducer)
   }
 
@@ -117,6 +121,8 @@ export class Tracim extends React.Component {
         i18n.changeLanguage(fetchGetUserIsConnected.json.lang)
 
         this.loadAppConfig()
+        this.loadLiveMessage()
+
         this.loadWorkspaceList()
         break
       case 401: props.dispatch(setUserConnected({ logged: false })); break
@@ -125,6 +131,7 @@ export class Tracim extends React.Component {
   }
 
   componentWillUnmount () {
+    this.liveMessageManager.closeLiveMessageConnection()
     document.removeEventListener(CUSTOM_EVENT.APP_CUSTOM_EVENT_LISTENER, this.customEventReducer)
   }
 
@@ -154,6 +161,21 @@ export class Tracim extends React.Component {
 
     const fetchGetContentTypeList = await props.dispatch(getContentTypeList())
     if (fetchGetContentTypeList.status === 200) props.dispatch(setContentTypeList(fetchGetContentTypeList.json))
+  }
+
+  loadLiveMessage = async () => {
+    const { props } = this
+    const response = await props.dispatch(getLiveMessage())
+
+    switch (response.status) {
+      case 200:
+        console.log('got live messages already sent')
+        break
+      default:
+        // props.dispatch(newFlashMessage(('Error while getting live messages'), 'error'))
+        console.log('error while getting live message')
+        break
+    }
   }
 
   loadWorkspaceList = async (openInSidebarId = undefined) => {
