+++ frontend/src/container/WorkspaceContent.jsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { connect } from 'react-redux'
 import { Link, withRouter, Route } from 'react-router-dom'
-import appFactory from '../appFactory.js'
-import i18n from '../i18n.js'
+import appFactory from '../util/appFactory.js'
+import i18n from '../util/i18n.js'
 import { translate } from 'react-i18next'
 import {
   PAGE,
@@ -10,7 +10,7 @@ import {
   CONTENT_TYPE,
   sortWorkspaceContents,
   SHARE_FOLDER_ID
-} from '../helper.js'
+} from '../util/helper.js'
 import Folder from '../component/Workspace/Folder.jsx'
 import ShareFolder from '../component/Workspace/ShareFolder.jsx'
 import ContentItem from '../component/Workspace/ContentItem.jsx'
