+++ frontend/src/reducer/currentWorkspace.js
@@ -1,5 +1,6 @@
 import {
   SET,
+  ADD,
   APPEND,
   REMOVE,
   WORKSPACE_DETAIL,
@@ -59,6 +60,21 @@ export default function currentWorkspace (state = defaultWorkspace, action) {
           isActive: m.is_active,
           doNotify: m.do_notify
         }))
+      }
+
+    case `${ADD}/${WORKSPACE_MEMBER}`:
+      return {
+        ...state,
+        memberList: [
+          ...state.memberList,
+          {
+            id: action.member.user_id,
+            publicName: action.member.user.public_name,
+            role: action.member.role.slug,
+            isActive: action.member.is_active,
+            doNotify: action.member.do_notify
+          }
+        ]
       }
 
     case `${SET}/${WORKSPACE_RECENT_ACTIVITY_LIST}`:
