+++ frontend/src/util/helper.js
@@ -1,9 +1,9 @@
 import i18n, { getBrowserLang } from './i18n.js'
 import { PROFILE_LIST, ROLE } from 'tracim_frontend_lib'
 
-const configEnv = process.env.NODE_ENV === 'test' ? require('../configEnv-test.json') : require('../configEnv.json')
+const configEnv = process.env.NODE_ENV === 'test' ? require('../../configEnv-test.json') : require('../../configEnv.json')
 
-const versionFile = require('./version.json')
+const versionFile = require('../version.json')
 export const TRACIM_APP_VERSION = versionFile.tracim_app_version
 export const SHARE_FOLDER_ID = -1
