+++ frontend/src/util/tinymceInit.js
@@ -1,4 +1,4 @@
-import i18n from '../i18n.js'
+import i18n from './i18n.js'
 
 (function () {
   function base64EncodeAndTinyMceInsert (files) {
